+++ extensions/universal-inbox/CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## [Unreleased]
 
+## [0.1.4] - 2024-03-13
+
 ### Added
 
 - Add Linear Project and Team icons
+- Add Slack notifications (ie. save for later) support
+- Set `Inbox` project as default when creating or planning a task
 
 ## [0.1.3] - 2024-02-05